requirements.txt
django==3.1.5
djangorestframework==3.11.0

pyjwt>=1.7.1